requirements.txt
flask
google-genai
gunicorn #sèvè pou render